test: rate limit
test not demonstrating correct behavior yet but results are nominally in
line with expectations. setting a "hard limit" that guarantees the
supervisor will never exceed a rate limit within a sliding window or
over the duration of a supervision will be more difficult and introduce
opportunties for bugs and slowness.